requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
django-rest==0.0.1
djangorestframework==3.10.3
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0